feat: implement CLI for asset management and CSV uploads
